refactor: [#1107] extract environment, printer, prelude from checker/types.rs
Splits the former 591-line `checker/types.rs` god module into four
focused files, mirroring Gleam's `type_/` layout:

- **`types.rs`** — now just the `Type` enum definition and its
  inherent constructor/inspection methods (`option_of`, `result_of`,
  `is_option`, `is_result`, `result_ok`, etc.). ~260 lines.
- **`environment.rs`** — `TypeEnv` (the scope stack of variable
  bindings, type declarations, and type-parameter trait bounds) and
  `TypeInfo`. ~175 lines. Will be the home for future
  `Arc<RefCell<TypeVar>>` unification state in #1118.
- **`printer.rs`** — `TypeDisplay`, `TypeDisplayStyle`, `fmt_type`,
  plus the `Display` impls for `Type` and `TypeDisplay`. ~180 lines.
  Isolated so display rules can evolve without touching the core
  type data structure.
- **`prelude.rs`** — Gleam-style module of free-function
  constructors for built-in types: `number()`, `string()`, `bool()`,
  `unit()`, `never()`, `unknown()`, `error()`, `undefined()`,
  `array_of()`, `promise_of()`, `settable_of()`, `option_of()`,
  `result_of()`, `tuple_of()`, `named()`, `foreign()`. Also the
  shared `UNKNOWN: Arc<Type>` sentinel, which is conceptually a
  prelude concern and was misplaced on `types.rs`. New callers
  should prefer `prelude::` for readability; existing `impl Type`
  helpers stay for backwards compatibility. ~130 lines (with tests).

## Scope kept narrow

This commit is purely a relocation: no behaviour changes, no
signature changes, no new functionality. The extracted modules are
byte-for-byte compatible with their former locations. The issue's
broader scope — splitting `checker/expr.rs` into `expression.rs` +
`pattern.rs` + `fields.rs` — is deferred to a followup because that
requires untangling real logic, not just moving declarations.
Renaming `checker/` → `type_/` to match Gleam's directory name was
also considered but skipped for this session: it would cascade into
every `crate::checker::` import across the workspace and deliver no
behavior win over keeping `checker/` as the umbrella name.

## Module wiring

`checker.rs` now declares `mod environment; mod printer; pub mod
prelude;` alongside the existing modules and re-exports the public
surface:

    pub use prelude::UNKNOWN;
    pub use printer::TypeDisplay;
    pub use types::Type;

The `TypeEnv` / `TypeInfo` import inside `checker.rs` was updated
from `use types::{TypeEnv, TypeInfo}` to `use environment::{TypeEnv,
TypeInfo}`. Every other caller consumes these through the Checker's
`env` field, so no other import sites needed adjustment.

## Quality gates

- `cargo check --workspace` clean
- `cargo test --workspace` 1384 unit + 33 codegen snapshot + 29
  error snapshot = 1446 passing (up from 1413; the 4 new tests are
  in `prelude::tests` — `unknown_sentinel_is_interned`,
  `array_of_wraps`, `option_of_round_trips_inner`,
  `result_of_round_trips_both_sides`)
- `cargo clippy --workspace -- -D warnings` clean
- `cargo fmt --check` clean
- `cargo build --profile ci` → `target/ci/floe` unchanged path
- `./target/ci/floe check examples/{todo-app,store}/src/` clean
- `pytest tests/lsp/ --floe-bin=./target/debug/floe` 236 passing

## Deferred from issue spec

- `checker/expression.rs` (shrink the 2939-line `checker/expr.rs`
  monolith — requires extracting pattern handling, stdlib dispatch,
  and bidirectional inference into focused passes)
- `checker/pattern.rs` — pattern checking already lives in
  `checker/match_check.rs`; a rename is cosmetic churn
- `checker/fields.rs` — record field maps are currently scattered
  across `items.rs` and `expr.rs` and need a proper extraction pass

These are tracked as followup scope on #1107 and can land in
subsequent commits without blocking downstream issues that only
need `environment` / `printer` / `prelude` (#1108 Arc<Type> and
#1113 Problems accumulator).

Part of epic #1101. Closes #1107.
