+++ crates/floe-core/src/checker/types.rs
@@ -1,14 +1,9 @@
-use std::collections::HashMap;
-use std::fmt;
-use std::sync::{Arc, LazyLock};
+//! The `Type` data structure and its inherent inspection/constructor
+//! methods. Ancillary concerns that used to live here (the scope
+//! stack, display formatting, prelude constructors) were split into
+//! `environment.rs`, `printer.rs`, and `prelude.rs` as part of #1107.
 
-use crate::parser::ast::TypeDef;
-
-/// Shared `Arc<Type::Unknown>` sentinel. Cloning bumps a refcount instead
-/// of allocating — the fallback path in `attach_types` hits this for every
-/// codegen-synthetic node and every post-error subtree, so keeping it
-/// interned matters for compiles with many errors.
-pub static UNKNOWN: LazyLock<Arc<Type>> = LazyLock::new(|| Arc::new(Type::Unknown));
+use super::printer::{TypeDisplay, TypeDisplayStyle};
 
 // ── Types ────────────────────────────────────────────────────────
 
@@ -254,338 +249,6 @@ impl Type {
         TypeDisplay {
             ty: self,
             style: TypeDisplayStyle::Stdlib,
-        }
-    }
-}
-
-// ── Type Display ────────────────────────────────────────────────
-
-/// Controls how types are formatted as strings.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum TypeDisplayStyle {
-    /// Default style used in error messages and LSP hover.
-    /// Type vars shown as `?T0`, tuples as `(a, b)`, optional params shown with `= None`.
-    Default,
-    /// Stdlib style used in stdlib hover/completion display.
-    /// Type vars shown as letter names (`T`, `U`, `V`), tuples as `[a, b]`.
-    Stdlib,
-}
-
-/// A wrapper for displaying a `Type` with a specific style.
-pub struct TypeDisplay<'a> {
-    ty: &'a Type,
-    style: TypeDisplayStyle,
-}
-
-/// Pretty-print a type variable index as a letter (0 -> T, 1 -> U, 2 -> V, ...).
-fn type_var_letter(index: usize) -> &'static str {
-    match index {
-        0 => "T",
-        1 => "U",
-        2 => "V",
-        3 => "W",
-        _ => "T",
-    }
-}
-
-/// Core type formatting logic shared by all display styles.
-fn fmt_type(ty: &Type, style: TypeDisplayStyle, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-    match ty {
-        Type::Number => f.write_str("number"),
-        Type::String => f.write_str("string"),
-        Type::Bool => f.write_str("boolean"),
-        Type::Undefined => f.write_str("undefined"),
-        Type::Named(n) | Type::Foreign(n) => f.write_str(n),
-        Type::Promise(inner) => {
-            write!(f, "Promise<{}>", TypeDisplay { ty: inner, style })
-        }
-        Type::Opaque { name, .. } => f.write_str(name),
-        Type::Settable(inner) => {
-            write!(f, "Settable<{}>", TypeDisplay { ty: inner, style })
-        }
-        Type::Function {
-            params,
-            required_params,
-            return_type,
-        } => {
-            f.write_str("(")?;
-            for (i, t) in params.iter().enumerate() {
-                if i > 0 {
-                    f.write_str(", ")?;
-                }
-                write!(f, "{}", TypeDisplay { ty: t, style })?;
-                if style == TypeDisplayStyle::Default
-                    && i >= *required_params
-                    && *required_params < params.len()
-                {
-                    f.write_str(" = None")?;
-                }
-            }
-            write!(
-                f,
-                ") -> {}",
-                TypeDisplay {
-                    ty: return_type,
-                    style
-                }
-            )
-        }
-        Type::Array(inner) => {
-            write!(f, "Array<{}>", TypeDisplay { ty: inner, style })
-        }
-        Type::Map { key, value } => {
-            write!(
-                f,
-                "Map<{}, {}>",
-                TypeDisplay { ty: key, style },
-                TypeDisplay { ty: value, style }
-            )
-        }
-        Type::RecordMap { key, value } => {
-            write!(
-                f,
-                "Record<{}, {}>",
-                TypeDisplay { ty: key, style },
-                TypeDisplay { ty: value, style }
-            )
-        }
-        Type::Set { element } => {
-            write!(f, "Set<{}>", TypeDisplay { ty: element, style })
-        }
-        Type::Tuple(types) => {
-            let (open, close) = match style {
-                TypeDisplayStyle::Stdlib => ("[", "]"),
-                TypeDisplayStyle::Default => ("(", ")"),
-            };
-            f.write_str(open)?;
-            for (i, t) in types.iter().enumerate() {
-                if i > 0 {
-                    f.write_str(", ")?;
-                }
-                write!(f, "{}", TypeDisplay { ty: t, style })?;
-            }
-            f.write_str(close)
-        }
-        Type::Record(fields) => {
-            f.write_str("{ ")?;
-            for (i, (n, t)) in fields.iter().enumerate() {
-                if i > 0 {
-                    f.write_str(", ")?;
-                }
-                write!(f, "{n}: {}", TypeDisplay { ty: t, style })?;
-            }
-            f.write_str(" }")
-        }
-        Type::Union { name, .. } => {
-            if name == crate::type_layout::TYPE_OPTION
-                && let Some(inner) = ty.option_inner()
-            {
-                return write!(f, "Option<{}>", TypeDisplay { ty: inner, style });
-            }
-            if name == crate::type_layout::TYPE_RESULT {
-                let ok_display = ty
-                    .result_ok()
-                    .map(|t| format!("{}", TypeDisplay { ty: t, style }))
-                    .unwrap_or_else(|| "unknown".to_string());
-                let err_display = ty
-                    .result_err()
-                    .map(|t| format!("{}", TypeDisplay { ty: t, style }))
-                    .unwrap_or_else(|| "unknown".to_string());
-                return write!(f, "Result<{ok_display}, {err_display}>");
-            }
-            f.write_str(name)
-        }
-        Type::TsUnion(members) => {
-            for (i, t) in members.iter().enumerate() {
-                if i > 0 {
-                    f.write_str(" | ")?;
-                }
-                write!(f, "{}", TypeDisplay { ty: t, style })?;
-            }
-            Ok(())
-        }
-        Type::StringLiteral(s) => write!(f, "\"{s}\""),
-        Type::Var(id) => match style {
-            TypeDisplayStyle::Stdlib => f.write_str(type_var_letter(*id)),
-            TypeDisplayStyle::Default => write!(f, "?T{id}"),
-        },
-        Type::Unknown => f.write_str("unknown"),
-        Type::Error => f.write_str("<error>"),
-        Type::Unit => f.write_str("()"),
-        Type::Never => f.write_str("never"),
-    }
-}
-
-impl fmt::Display for Type {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        fmt_type(self, TypeDisplayStyle::Default, f)
-    }
-}
-
-impl fmt::Display for TypeDisplay<'_> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        fmt_type(self.ty, self.style, f)
-    }
-}
-
-// ── Type Environment ─────────────────────────────────────────────
-
-/// Tracks types of variables, functions, and type declarations in scope.
-#[derive(Debug, Clone)]
-pub(crate) struct TypeEnv {
-    /// Stack of scopes (innermost last). Each scope maps names to types.
-    pub(crate) scopes: Vec<HashMap<String, Type>>,
-    /// Type declarations: type name -> TypeDef + metadata
-    type_defs: HashMap<String, TypeInfo>,
-    /// Trait bounds on type parameters: param name -> [trait names]
-    type_param_bounds: HashMap<String, Vec<String>>,
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct TypeInfo {
-    #[allow(dead_code)]
-    pub(crate) def: TypeDef,
-    pub(crate) opaque: bool,
-    #[allow(dead_code)]
-    pub(crate) type_params: Vec<String>,
-}
-
-impl TypeEnv {
-    pub(crate) fn new() -> Self {
-        Self {
-            scopes: vec![HashMap::new()],
-            type_defs: HashMap::new(),
-            type_param_bounds: HashMap::new(),
-        }
-    }
-
-    pub(crate) fn push_scope(&mut self) {
-        self.scopes.push(HashMap::new());
-    }
-
-    pub(crate) fn pop_scope(&mut self) {
-        self.scopes.pop();
-    }
-
-    pub(crate) fn define(&mut self, name: &str, ty: Type) {
-        if let Some(scope) = self.scopes.last_mut() {
-            scope.insert(name.to_string(), ty);
-        }
-    }
-
-    /// Define a name in the parent scope (second-to-last), used to update
-    /// function types after inferring the return type from the body.
-    pub(crate) fn define_in_parent_scope(&mut self, name: &str, ty: Type) {
-        let len = self.scopes.len();
-        if len >= 2 {
-            self.scopes[len - 2].insert(name.to_string(), ty);
-        }
-    }
-
-    /// Check if a name is defined in the current (innermost) scope only.
-    pub(crate) fn is_defined_in_current_scope(&self, name: &str) -> bool {
-        self.scopes
-            .last()
-            .is_some_and(|scope| scope.contains_key(name))
-    }
-
-    pub(crate) fn lookup(&self, name: &str) -> Option<&Type> {
-        for scope in self.scopes.iter().rev() {
-            if let Some(ty) = scope.get(name) {
-                return Some(ty);
-            }
-        }
-        None
-    }
-
-    pub(crate) fn define_type(&mut self, name: &str, info: TypeInfo) {
-        self.type_defs.insert(name.to_string(), info);
-    }
-
-    pub(crate) fn lookup_type(&self, name: &str) -> Option<&TypeInfo> {
-        self.type_defs.get(name)
-    }
-
-    /// Define trait bounds for a type parameter.
-    pub(crate) fn define_type_param_bounds(&mut self, name: &str, bounds: Vec<String>) {
-        self.type_param_bounds.insert(name.to_string(), bounds);
-    }
-
-    /// Look up trait bounds for a type parameter.
-    pub(crate) fn get_type_param_bounds(&self, name: &str) -> Option<&Vec<String>> {
-        self.type_param_bounds.get(name)
-    }
-
-    /// Resolve a `Type::Named("Foo")` to its concrete type by looking up the type definition.
-    /// For records, returns `Type::Record(fields)`. For unions, returns `Type::Union`.
-    /// For aliases, follows the chain. Primitives and non-Named types pass through unchanged.
-    pub(crate) fn resolve_to_concrete(
-        &self,
-        ty: &Type,
-        resolve_type_fn: &dyn Fn(&crate::parser::ast::TypeExpr) -> Type,
-    ) -> Type {
-        match ty {
-            Type::Named(name) | Type::Foreign(name) => {
-                if let Some(info) = self.lookup_type(name) {
-                    self.resolve_type_def(name, info, resolve_type_fn)
-                } else {
-                    ty.clone()
-                }
-            }
-            Type::Promise(_) => ty.clone(),
-            _ => ty.clone(),
-        }
-    }
-
-    fn resolve_type_def(
-        &self,
-        name: &str,
-        info: &TypeInfo,
-        resolve_type_fn: &dyn Fn(&crate::parser::ast::TypeExpr) -> Type,
-    ) -> Type {
-        match &info.def {
-            crate::parser::ast::TypeDef::Record(entries) => {
-                let field_types: Vec<_> = entries
-                    .iter()
-                    .filter_map(|e| e.as_field())
-                    .map(|f| (f.name.clone(), resolve_type_fn(&f.type_ann)))
-                    .collect();
-                Type::Record(field_types)
-            }
-            crate::parser::ast::TypeDef::Union(variants) => {
-                let var_types: Vec<_> = variants
-                    .iter()
-                    .map(|v| {
-                        let field_types: Vec<_> = v
-                            .fields
-                            .iter()
-                            .map(|f| resolve_type_fn(&f.type_ann))
-                            .collect();
-                        (v.name.clone(), field_types)
-                    })
-                    .collect();
-                Type::Union {
-                    name: name.to_string(),
-                    variants: var_types,
-                }
-            }
-            crate::parser::ast::TypeDef::StringLiteralUnion(variants) => Type::TsUnion(
-                variants
-                    .iter()
-                    .map(|s| Type::StringLiteral(s.clone()))
-                    .collect(),
-            ),
-            crate::parser::ast::TypeDef::Alias(type_expr) => {
-                // For typeof aliases, use the pre-resolved env binding
-                // (simple_resolve_type_expr can't resolve typeof without env context)
-                if matches!(type_expr.kind, crate::parser::ast::TypeExprKind::TypeOf(_))
-                    && let Some(resolved) = self.lookup(name)
-                {
-                    return self.resolve_to_concrete(&resolved.clone(), resolve_type_fn);
-                }
-                let resolved = resolve_type_fn(type_expr);
-                self.resolve_to_concrete(&resolved, resolve_type_fn)
-            }
         }
     }
 }
